+++ src/Moq/Invocation.cs
@@ -1,6 +1,7 @@
 // Copyright (c) 2007, Clarius Consulting, Manas Technology Solutions, InSTEDD.
 // All rights reserved. Licensed under the BSD 3-Clause License; see License.txt.
 
+using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Reflection;
@@ -12,27 +13,45 @@ namespace Moq
 	{
 		private object[] arguments;
 		private MethodInfo method;
+		private MethodInfo methodImplementation;
+		private readonly Type proxyType;
 		private object returnValue;
 		private VerificationState verificationState;
 
 		/// <summary>
 		/// Initializes a new instance of the <see cref="Invocation"/> class.
 		/// </summary>
+		/// <param name="proxyType">The <see cref="Type"/> of the concrete proxy object on which a method is being invoked.</param>
 		/// <param name="method">The method being invoked.</param>
 		/// <param name="arguments">The arguments with which the specified <paramref name="method"/> is being invoked.</param>
-		protected Invocation(MethodInfo method, params object[] arguments)
+		protected Invocation(Type proxyType, MethodInfo method, params object[] arguments)
 		{
+			Debug.Assert(proxyType != null);
 			Debug.Assert(arguments != null);
 			Debug.Assert(method != null);
 
 			this.arguments = arguments;
 			this.method = method;
+			this.proxyType = proxyType;
 		}
 
 		/// <summary>
 		/// Gets the method of the invocation.
 		/// </summary>
 		public MethodInfo Method => this.method;
+
+		public MethodInfo MethodImplementation
+		{
+			get
+			{
+				if (this.methodImplementation == null)
+				{
+					this.methodImplementation = this.method.GetImplementingMethod(this.proxyType);
+				}
+
+				return this.methodImplementation;
+			}
+		}
 
 		/// <summary>
 		/// Gets the arguments of the invocation.
@@ -44,6 +63,8 @@ namespace Moq
 		public object[] Arguments => this.arguments;
 
 		IReadOnlyList<object> IInvocation.Arguments => this.arguments;
+
+		public Type ProxyType => this.proxyType;
 
 		public object ReturnValue => this.returnValue;