+++ CHANGELOG.md
@@ -19,6 +19,8 @@ The format is loosely based on [Keep a Changelog](http://keepachangelog.com/en/1
 
 * Moq will throw when it detects that an argument matcher will never match anything due to the presence of an implicit conversion operator. (@michelcedric, #897, #898)
 
+* New algorithm for matching invoked methods against methods specified in setup/verification expressions. (@stakx, #904)
+
 #### Added
 
 * Added support for setup and verification of the event handlers through `Setup[Add|Remove]` and `Verify[Add|Remove|All]` (@lepijohnny, #825) 
@@ -26,6 +28,8 @@ The format is loosely based on [Keep a Changelog](http://keepachangelog.com/en/1
 * Added support for lambda expressions while creating a mock through `new Mock<SomeType>(() => new SomeType("a", "b"))` and `repository.Create<SomeType>(() => new SomeType("a", "b"))`. This makes the process of mocking a class without a parameterless constructor simpler (compiler syntax checker...). (@frblondin, #884)
 
 #### Fixed
+
+* Moq does not mock explicit interface implementation and `protected virtual` correctly. (@oddbear, #657)
 
 * `Invocations.Clear()` does not cause `Verify` to fail (@jchessir, #733)
 
@@ -36,6 +40,8 @@ The format is loosely based on [Keep a Changelog](http://keepachangelog.com/en/1
 * `Verify` throws `TargetInvocationException` instead of `MockException` when one of the recorded invocations was to an async method that threw. (@Cufeadir, #883)
 
 * Regression in 4.12.0: `SetupAllProperties` removes indexer setups. (@stakx, #901)
+
+* Parameter types are ignored when matching an invoked generic method against setups. (@stakx, #903)
 
 
 ## 4.12.0 (2019-06-20)